Solution to Check Less Than

+++ Check Less Than@@67882a731b4392d8fc1577d7/code.c
@@ -1,6 +1,6 @@
 void main() {
-    int a,b;
-    scanf("%i %i",&a,&b);
+    float a,b;
+    scanf("%f %f",&a,&b);
     if(a<b)
     {
     printf("True");
